Refactor navbar: extract NavigationUserMenu
