+++ frontend/src/layouts/NavigationDrawer.tsx
@@ -1,3 +1,4 @@
+// See MUI Drawer Navbar example
 import React from "react";
 import { NavLink } from "react-router-dom";
